+++ src/app/api/admin/posts/[id]/route.ts
@@ -1,6 +1,7 @@
 import prisma from "@/lib/prisma";
 import { NextResponse, NextRequest } from "next/server";
 import { Post } from "@prisma/client";
+import { supabase } from "@/utils/supabase";
 
 type RouteParams = {
   params: {
@@ -11,7 +12,7 @@ type RouteParams = {
 type RequestBody = {
   title: string;
   content: string;
-  coverImageURL: string;
+  coverImageKey: string; // 変更
   categoryIds: string[];
 };
 
@@ -21,7 +22,7 @@ export const PUT = async (req: NextRequest, routeParams: RouteParams) => {
     const requestBody: RequestBody = await req.json();
 
     // 分割代入
-    const { title, content, coverImageURL, categoryIds } = requestBody;
+    const { title, content, coverImageKey, categoryIds } = requestBody; // 変更
 
     // categoryIds に該当するカテゴリが存在するか確認
     const categories = await prisma.category.findMany({
@@ -46,7 +47,7 @@ export const PUT = async (req: NextRequest, routeParams: RouteParams) => {
       data: {
         title, // title: title の省略形であることに注意。以下も同様
         content,
-        coverImageURL,
+        coverImageKey, // 変更
       },
     });
 
@@ -73,6 +74,32 @@ export const PUT = async (req: NextRequest, routeParams: RouteParams) => {
 export const DELETE = async (req: NextRequest, routeParams: RouteParams) => {
   try {
     const id = routeParams.params.id;
+
+    // 認証チェック
+    const token = req.headers.get("Authorization") ?? "";
+    const { data, error } = await supabase.auth.getUser(token);
+    if (error || !data.user) {
+      return NextResponse.json(
+        { error: "認証に失敗しました" },
+        { status: 401 }
+      );
+    }
+
+    // 管理者チェック
+    const { user } = data;
+    const { data: userData, error: userError } = await supabase
+      .from("profiles")
+      .select("is_admin")
+      .eq("id", user.id)
+      .single();
+
+    if (userError || !userData.is_admin) {
+      return NextResponse.json(
+        { error: "管理者権限が必要です" },
+        { status: 403 }
+      );
+    }
+
     const post: Post = await prisma.post.delete({
       where: { id },
     });
